File clean up and rename few files
