--- src/features/RelatedItemsAndComparison/OutfitList/OutfitItemCard/OutfitItemCard.module.css
@@ -1,37 +1,0 @@
-.product_card {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  border:1px solid black;
-  width: 250px;
-  /* margin: 0px auto; */
-  cursor: pointer;
-}
-
-.product_card:hover, .product_card:active {
-  box-shadow: 6px, 6px, 12px, rgba(0, 0, 0, 0.5);
-}
-
-.product_card_header {
-  /* FIXME: */
-  position: relative;
-  flex:1;
-  color: black;
-}
-
-.product_img {
-  /* FIXME: */
-  color: black;
-}
-
-.product_action_button {
-  /* FIXME: */
-  font-size: 36px;
-  color: red;
-  position: absolute;
-  top: -4px;
-  right: 6px;
-}
-.product_card_body {
-  flex:1;
-}
