+++ src/features/RelatedItemsAndComparison/OutfitList/OutfitList.jsx
@@ -1,7 +1,3 @@
-/* eslint-disable implicit-arrow-linebreak */
-/* eslint-disable react/jsx-wrap-multilines */
-/* eslint-disable jsx-a11y/click-events-have-key-events */
-/* eslint-disable jsx-a11y/no-noninteractive-element-interactions */
 /* eslint-disable import/extensions */
 import React, { useState } from 'react';
 import { useAppDispatch, useAppSelector } from '../../../app/redux-hooks';
@@ -9,7 +5,7 @@ import PropTypes from 'prop-types';
 import ListContainer from '../helpers/ListContainer/ListContainer.jsx';
 import { addOutfit, deleteOutfit } from './outfitListSlice.jsx';
 import Button from '../../UI/Button.jsx';
-import OutfitItemCard from './OutfitItemCard/OutfitItemCard.jsx';
+import OutfitListItem from './OutfitListItem.jsx';
 // import styles from './OutfitList.module.css';
 
 export default function OutfitList({ currentViewItemId }) {
@@ -34,8 +30,8 @@ export default function OutfitList({ currentViewItemId }) {
       <p>My Outfit List!!!</p>
       <ListContainer>
         {!isCurrentItemAdded && <Button onClick={() => handleAddOutfit()}>Add to List</Button>}
-        { userOutfitList?.map((outfitId) =>
-          <OutfitItemCard
+        {userOutfitList?.map((outfitId) =>
+          <OutfitListItem
             key={outfitId}
             productId={outfitId}
             handleDeleteOutfit={handleDeleteOutfit.bind(null, outfitId)}
