+++ src/features/RelatedItemsAndComparison/RelatedProductsList/RelatedProductItem.module.css
